requirements.txt
djangorestframework
environs
bcrypt
Django~=5.0.6